pyproject.toml
[project]
name = "prediction-markets"
version = "0.1.0"
description = "Core wrapper library for unified prediction market access"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    # HTTP/WebSocket
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "requests>=2.31.0",
    # Config
    "python-dotenv>=1.0.0",
    # Ethereum/Web3
    "eth-account>=0.11.0",
    "eth-abi>=5.0.0",
    "web3>=6.0.0",
    # Polymarket SDK
    "py-order-utils>=0.1.0",
    "py-clob-client>=0.17.0",
    "py-builder-relayer-client>=0.0.1",
    "py-builder-signing-sdk>=0.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "aioresponses>=0.7.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prediction_markets"]
